% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{dijkstra}
\source{
See more at: \url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{a graph as data frame.}

\item{init_node}{the initial node.}
}
\value{
The shortest path of \code{graph} with the initial node \code{init_node}
}
\description{
The algorithm calculates the shortest path of a graph \code{wiki_graph} with an initial node.
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3)
}
